+++ tests/Integration/Bridge/Laravel/RegexParserServiceProviderTest.php
@@ -90,7 +90,7 @@ final class RegexParserServiceProviderTest extends TestCase
         $regex->parse('/abc/');
 
         $cache = new FilesystemCache($cacheDir);
-        $cacheSeed = "/abc/\n#cache=".Regex::CACHE_VERSION;
+        $cacheSeed = "/abc/\n#cache=".Regex::CACHE_VERSION."\n#php_version=".\PHP_VERSION_ID;
         $cacheFile = $cache->generateKey($cacheSeed);
 
         $this->assertFileExists($cacheFile);